Move row_object class to replication_object.[ch]pp
