+++ src/transaction/log_impl.h
@@ -57,11 +57,7 @@
 #include "thread_entry.hpp"
 #include "transaction_transient.hpp"
 #include "log_generator.hpp"
-
-namespace cubreplication
-{
-  class copy_context;
-}
+#include "replication_db_copy.hpp"
 
 #include <assert.h>
 #if defined(SOLARIS)
@@ -545,7 +541,7 @@ struct log_tdes
   // *INDENT-OFF*
 #if defined (SERVER_MODE) || (defined (SA_MODE) && defined (__cplusplus))
   cubreplication::log_generator replication_log_generator;
-  cubreplication::copy_context *replication_copy_context;
+  cubreplication::copy_context replication_copy_context;
 
   bool is_active_worker_transaction () const;
   bool is_system_transaction () const;
